Create web-monitor.sh
I once used this script in conjunction with a cron entry to check if my site was up or not — perhaps nginx or PHP had crashed for some peculiar reason, and needed to be restarted, but I wasn't online at the time? Well, that's where this script came in handy.
It worked fine at the time and it still does. However, there are far better ways of handling issues such as this. I am however including it for posterity.